Check id of messages when loading messages.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@3022 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qm/src/model/account.cpp
@@ -2715,7 +2715,7 @@ MessageHolder* qm::Account::storeMessage(NormalFolder* pFolder,
 		pTime = &time;
 	}
 	
-	if (nId == static_cast<unsigned int>(-1)) {
+	if (nId == -1) {
 		nId = pFolder->generateId();
 		if (nId == -1)
 			return 0;
